fix(print): wrap headers by display width
